function that returns all elems of a list

+++ 0x12-singly_linked_lists/lists.h
@@ -10,7 +10,7 @@
  */
 struct Node
 {
-	unsigned long int length;
+	unsigned int length;
 	char *str;
 	struct Node *next;
 };
